EchoSafe/backend/requirements.txt
fastapi==0.115.6
uvicorn==0.32.1
pymongo==4.10.1
bcrypt==4.2.1
PyJWT==2.10.1
